requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
anthropic==0.49.0
pdf2image==1.17.0
python-dotenv==1.1.0
pyyaml==6.0.2
python-multipart==0.0.20
httpx==0.28.1
openai==1.75.0
Pillow==11.2.1
pytest==8.3.5
starlette[testclient]==0.46.2
ruff==0.11.7
